Split source into *.h & *.cpp
