+++ src/LibXMLUtil.h
@@ -1,0 +1,6 @@
+#ifndef LIBXMLUTIL_H
+#define LIBXMLUTIL_H
+
+xmlNodePtr findFirst(xmlNodePtr node, const char* xpathExpr, xmlXPathContextPtr xpathCtxt);
+
+#endif /* !LIBXMLUTIL_H */
