+++ src/TypeAnalyzer.h
@@ -1,0 +1,7 @@
+#ifndef TYPEANALYZER_H
+#define TYPEANALYZER_H
+
+using TypeAnalyzer = Reality<TypeMap&>;
+TypeMap parseTypeTable(xmlDocPtr doc);
+
+#endif /* !TYPEANALYZER_H */
